refactor: ImageType enum을 public → package-private으로 변경하여 외부 노출 방지
